# backend/docs/swagger.yaml
definitions:
...
  models.Housekeeping:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: integer
      room:
        $ref: '#/definitions/models.Room'
      room_id:
        type: integer
      staff_id:
        description: 清洁人员 ID
        type: integer
      start_time:
        type: string
      status:
        description: 状态：pending, in_progress, completed
        type: string
      type:
        description: 类型：daily(续住清), checkout(退房清), deep(深清)
        type: string
      updated_at:
        type: string
    type: object
...
  models.Maintenance:
    properties:
      created_at:
        type: string
      description:
        description: 问题描述
        type: string
      id:
        type: integer
      remark:
        description: 维修备注
        type: string
      reporter_id:
        description: 报修人 (员工 ID)
        type: integer
      room:
        $ref: '#/definitions/models.Room'
      room_id:
        type: integer
      status:
        description: 状态：pending, in_progress, completed, cancelled
        type: string
      type:
        description: 报修类型：plumbing, electrical, furniture, etc.
        type: string
      updated_at:
        type: string
      worker_id:
        description: 维修人 (员工 ID)
        type: integer
    type: object
...
  models.Room:
    properties:
      area:
        description: 面积（平方米）
        type: number
      bed_type:
        description: 床型：单人床、双人床、大床
        type: string
      capacity:
        description: 可住人数
        type: integer
      clean_status:
        description: 清洁状态：clean(干净), dirty(脏房), inspecting(待查)
        type: string
      created_at:
        description: 创建时间
        type: string
      description:
        description: 房间描述
        type: string
      facilities:
        description: 设施（JSON 字符串）
        type: string
      floor:
        description: 楼层
        type: integer
      height:
        description: 高度
        type: integer
      id:
        description: 主键
        type: integer
      images:
        description: 图片 URL（JSON 数组）
        type: string
      left:
        description: 左边界
        type: integer
      original_price:
        description: 原价
        type: number
      price:
        description: 价格（每晚）
        type: number
      room_number:
        description: 房间号（唯一，有索引）
        type: string
      room_type:
        description: 房间类型（有索引）
        type: string
      status:
        description: 状态：available(空闲), occupied(在住), maintenance(维修)
        type: string
      top:
        description: 上边界
        type: integer
      updated_at:
        description: 更新时间
        type: string
      width:
        description: 宽度
        type: integer
    type: object
...
  utils.ErrorInfo:
    properties:
      code:
        description: 错误代码
        type: string
      message:
        description: 错误消息
        type: string
    type: object
...
  utils.Response:
    properties:
      data:
        description: 返回的数据
      error:
        allOf:
        - $ref: '#/definitions/utils.ErrorInfo'
        description: 错误信息
      message:
        description: 提示消息
        type: string
      success:
        description: 是否成功
        type: boolean
    type: object
host: 127.0.0.1:18080
info:
  contact:
    email: support@gohotel.com
    name: API Support
  description: 这是一个酒店预订管理系统的 RESTful API 服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 酒店管理系统 API
  version: "1.0"
paths:
...
  /api/admin/work-orders/cleaning:
    post:
      consumes:
      - application/json
      description: 创建新的清洁任务
      parameters:
      - description: '创建清洁任务 {room_id: int, type: string}'
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 创建清洁任务
      tags:
      - 工单管理
  /api/admin/work-orders/cleaning/{id}/assign:
    post:
      consumes:
      - application/json
      description: 为工单分派负责员工
      parameters:
      - description: 工单ID
        in: path
        name: id
        required: true
        type: integer
      - description: '分派请求 {staff_id: int}'
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 分派成功
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 分派员工
      tags:
      - 工单管理
  /api/admin/work-orders/cleaning/{id}/complete:
    post:
      consumes:
      - application/json
      description: 标记清洁任务为已完成
      parameters:
      - description: 清洁任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 完成清洁任务
      tags:
      - 工单管理
  /api/admin/work-orders/cleanings:
    get:
      consumes:
      - application/json
      description: 获取所有清洁任务，支持按状态筛选
      parameters:
      - description: 状态筛选
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Housekeeping'
                  type: array
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 获取清洁任务列表
      tags:
      - 工单管理
  /api/admin/work-orders/repair:
    post:
      consumes:
      - application/json
      description: 创建新的维修请求
      parameters:
      - description: '创建维修请求 {room_id: int, type: string, description: string}'
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 创建维修请求
      tags:
      - 工单管理
  /api/admin/work-orders/repair/{id}/complete:
    post:
      consumes:
      - application/json
      description: 标记维修任务为已完成
      parameters:
      - description: 维修任务ID
        in: path
        name: id
        required: true
        type: integer
      - description: '备注 {remark: string}'
        in: body
        name: request
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 完成维修任务
      tags:
      - 工单管理
  /api/admin/work-orders/repairs:
    get:
      consumes:
      - application/json
      description: 获取所有维修任务，支持按状态筛选
      parameters:
      - description: 状态筛选
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Maintenance'
                  type: array
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 获取维修任务列表
      tags:
      - 工单管理
...
securityDefinitions:
  Bearer:
    description: 输入 "Bearer {token}" 格式的 JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"